+++ core/src/main/java/io/fabric8/maven/core/extenvvar/JsonSchemaProperty.java
@@ -14,13 +14,11 @@
  * permissions and limitations under the License.
  */
 
-package io.fabric8.maven.core.support;
+package io.fabric8.maven.core.extenvvar;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-/**
- */
-public class JsonSchemaProperty {
+class JsonSchemaProperty {
     private String type = "string";
     private String format;
     @JsonProperty("default")
