+++ core/src/main/java/io/fabric8/maven/core/handler/ProbeHandler.java
@@ -22,7 +22,7 @@ import java.util.List;
 
 import io.fabric8.kubernetes.api.model.*;
 import io.fabric8.maven.core.config.ProbeConfiguration;
-import io.fabric8.maven.core.support.Commandline;
+import io.fabric8.maven.core.util.Commandline;
 import io.fabric8.utils.Strings;
 
 /**
