+++ enricher/api/src/main/java/io/fabric8/maven/enricher/api/BaseEnricher.java
@@ -67,7 +67,7 @@ public abstract class BaseEnricher implements Enricher {
         return config.get(key, defaultVal);
     }
 
-    protected EnricherContext getBuildContext() {
+    protected EnricherContext getContext() {
         return buildContext;
     }
